requirements.txt
langchain==0.3.25
openai==0.27.6
PyPDF2==3.0.1
python-dotenv==1.1.0
streamlit==1.18.1
transformers==4.31.0
